requirements.txt
django-cors-headers=4.1.0
django=4.2.2
djangorestframework=3.14